+++ tests/unit_tests/AuthenticationTest.php
@@ -1,22 +1,35 @@
 <?php
 namespace Payplug;
 use Payplug;
+use Payplug\Core\HttpClient;
 
 /**
 * @group unit
 * @group ci
 * @group recommended
 */
-class AuthenticationTest extends \PHPUnit_Framework_TestCase
+class AuthenticationTest extends \PHPUnit\Framework\TestCase
 {
     private $_requestMock;
 
-    protected function setUp()
+    /**
+     * @before
+     */
+    protected function setUpTest()
     {
-        $this->_configuration = new Payplug\Payplug('abc');
+        $this->_configuration = new \Payplug\Payplug('abc');
         Payplug\Payplug::setDefaultConfiguration($this->_configuration);
 
-        $this->_requestMock = $this->getMock('\Payplug\Core\IHttpRequest');
+        $this->_requestMock = $this->createMock('\Payplug\Core\IHttpRequest');
+        Core\HttpClient::$REQUEST_HANDLER = $this->_requestMock;
+    }
+
+    protected function setUpTwice()
+    {
+        $this->_configuration = new \Payplug\Payplug('abc','1970-01-01');
+        Payplug\Payplug::setDefaultConfiguration($this->_configuration);
+
+        $this->_requestMock = $this->createMock('\Payplug\Core\IHttpRequest');
         Core\HttpClient::$REQUEST_HANDLER = $this->_requestMock;
     }
 
@@ -105,7 +118,10 @@ class AuthenticationTest extends \PHPUnit_Framework_TestCase
                 'max_amounts' => array(),
             ),
             'permissions' => array(
-                'use_live_mode' => true,
+                'can_use_oney' => true,
+                'use_live_mode' => false,
+                'can_create_deferred_payment' => true,
+                'can_create_installment_plan' => false,
                 'can_save_cards' => false,
             ),
         );
@@ -127,7 +143,10 @@ class AuthenticationTest extends \PHPUnit_Framework_TestCase
 
         $permissions = Authentication::getPermissions($this->_configuration);
 
-        $this->assertEquals(true, $permissions['use_live_mode']);
+        $this->assertEquals(true, $permissions['can_use_oney']);
+        $this->assertEquals(false, $permissions['use_live_mode']);
+        $this->assertEquals(true, $permissions['can_create_deferred_payment']);
+        $this->assertEquals(false, $permissions['can_create_installment_plan']);
         $this->assertEquals(false, $permissions['can_save_cards']);
     }
 }