+++ tests/unit_tests/Resource/PaymentTest.php
@@ -7,17 +7,28 @@ use Payplug;
  * @group ci
  * @group recommended
  */
-class PaymentTest extends \PHPUnit_Framework_TestCase
+class PaymentTest extends \PHPUnit\Framework\TestCase
 {
     private $_requestMock;
     private $_configuration;
 
-    protected function setUp()
+    /**
+  * @before
+  */ protected function setUpTest()
     {
         $this->_configuration = new Payplug\Payplug('abc');
         Payplug\Payplug::setDefaultConfiguration($this->_configuration);
 
-        $this->_requestMock = $this->getMock('\Payplug\Core\IHttpRequest');
+        $this->_requestMock = $this->createMock('\Payplug\Core\IHttpRequest');
+        Payplug\Core\HttpClient::$REQUEST_HANDLER = $this->_requestMock;
+    }
+
+    protected function setUpTwice()
+    {
+        $this->_configuration = new Payplug\Payplug('abc','1970-01-01');
+        Payplug\Payplug::setDefaultConfiguration($this->_configuration);
+
+        $this->_requestMock = $this->createMock('\Payplug\Core\IHttpRequest');
         Payplug\Core\HttpClient::$REQUEST_HANDLER = $this->_requestMock;
     }
 
@@ -542,6 +553,16 @@ class PaymentTest extends \PHPUnit_Framework_TestCase
         unset($GLOBALS['CURLOPT_URL_DATA']);
     }
 
+    public function testPaymentRetrieveWithInvalidId() {
+        $this->expectException('\PayPlug\Exception\UndefinedAttributeException');
+        Payment::retrieve('');
+    }
+
+    public function testPaymentRetrieveWithIdIsNull() {
+        $this->expectException('\PayPlug\Exception\UndefinedAttributeException');
+        Payment::retrieve(NULL);
+    }
+
     public function testPaymentList()
     {
         $GLOBALS['CURLOPT_URL_DATA'] = null;
@@ -605,7 +626,7 @@ class PaymentTest extends \PHPUnit_Framework_TestCase
 
     public function testPaymentRefundWhenPaymentIsInvalid()
     {
-        $this->setExpectedException('\PayPlug\Exception\InvalidPaymentException');
+        $this->expectException('\PayPlug\Exception\InvalidPaymentException');
 
         $payment = Payment::fromAttributes(array('fake' => 'payment'));
         $payment->refund(array('amount' => 3300));
@@ -652,7 +673,7 @@ class PaymentTest extends \PHPUnit_Framework_TestCase
 
     public function testPaymentListRefundsWhenPaymentIsInvalid()
     {
-        $this->setExpectedException('\PayPlug\Exception\InvalidPaymentException');
+        $this->expectException('\PayPlug\Exception\InvalidPaymentException');
 
         $payment = Payment::fromAttributes(array('fake' => 'payment'));
         $payment->listRefunds();
@@ -704,7 +725,7 @@ class PaymentTest extends \PHPUnit_Framework_TestCase
 
     public function testRetrieveConsistentPaymentWhenIdIsUndefined()
     {
-        $this->setExpectedException('\PayPlug\Exception\UndefinedAttributeException');
+        $this->expectException('\PayPlug\Exception\UndefinedAttributeException');
 
         $payment = Payment::fromAttributes(array('this_payment' => 'has_no_id'));
         $payment->getConsistentResource();